+++ src/apiCalls.js
@@ -9,21 +9,25 @@ let ingredientsData = [];
 let recipeData = [];
 
 
-export function fetchUsersData = () => {
+export function fetchUsersData() {
     fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/users')
-    .then(response => response.json());
-    .then(data => usersData.push(data));
+    .then(response => response.json())
+    .then(data => usersData.push(data))
 
 }
-export function fetchIngredientsData = () => {
+export function fetchIngredientsData() {
   fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/ingredients')
-  .then(response => response.json());
-  .then(data => ingredientsData.push(data));
+  .then(response => response.json())
+  .then(data => ingredientsData.push(data))
 }
-export function fetchRecipeData = () => {
+export function fetchRecipeData() {
   fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/recipes')
-  .then(response => response.json());
-  .then(data => recipeData.push(data));
+  .then(response => response.json())
+  .then(data => recipeData.push(data))
 }
 
-export {usersData, ingredientsData, recipeData};
+export default {
+usersData,
+ingredientsData,
+recipeData
+}
